added hooks to form, fixed css
