+++ src/pages/SingleGame.js
@@ -3,13 +3,14 @@ import GameInfoForm from "../components/CreateGameForm";
 import { PageWrapper } from "../components/ComponentStyles";
 import Header from "../containers/Header";
 import { useQuery } from "@apollo/client";
-import { MATCH_BY_ID, PLAYERS_BY_MATCH_ID } from "../queries";
+import { MATCH_BY_ID, PLAYERS_BY_MATCH_ID, CURRENT_USER } from "../queries";
 
 const SingleGame = () => {
+  const currentUser = useQuery(CURRENT_USER);
   const matchById = useQuery(MATCH_BY_ID, {
     variables: { id: window.location.pathname.slice(13) },
   });
-
+  
   const playersByMatchId = useQuery(PLAYERS_BY_MATCH_ID, {
     variables: {
       id: window.location.pathname.slice(13),
@@ -80,7 +81,12 @@ const SingleGame = () => {
     <PageWrapper>
       <Header />
       <GameInfoContainer title="Pelaajan Peli">
-        <GameInfoForm matchData={matchData} singleGameView={true} />
+        <GameInfoForm 
+          matchData={matchData} 
+          singleGameView={true} 
+          editMode={true}
+          currentUser={currentUser}
+          />
       </GameInfoContainer>
     </PageWrapper>
   );
